refactor: extract Chrome API helpers to src/lib/storage.ts
