+++ src/sidepanel/components/App.tsx
@@ -2,6 +2,7 @@ import { useState, useCallback, useRef, useEffect } from "react";
 import type { StorageType, StorageEntry, StorageChangeEvent, StorageChangePortMessage } from "@/shared/types";
 
 import { filterEntries } from "@/lib/filter";
+import { getActiveTabId, readStorage, writeStorage, removeFromStorage, importToStorage } from "@/lib/storage";
 import styles from "./App.module.css";
 import { StorageToggle } from "./StorageToggle";
 import { SearchBar } from "./SearchBar";
@@ -11,48 +12,6 @@ import { ImportExport } from "./ImportExport";
 import { ChangeLog } from "./ChangeLog";
 
 type LoadState = "idle" | "loading" | "ready" | "error";
-
-async function getActiveTabId(): Promise<number | null> {
-  const [tab] = await chrome.tabs.query({ active: true, currentWindow: true });
-  return tab?.id ?? null;
-}
-
-function readStorage(storageType: string): Array<{ key: string; value: string }> {
-  const storage = storageType === "localStorage" ? localStorage : sessionStorage;
-  const entries: Array<{ key: string; value: string }> = [];
-  for (let i = 0; i < storage.length; i++) {
-    const key = storage.key(i);
-    if (key !== null) {
-      entries.push({ key, value: storage.getItem(key) ?? "" });
-    }
-  }
-  return entries;
-}
-
-function writeStorage(storageType: string, key: string, value: string): void {
-  const storage = storageType === "localStorage" ? localStorage : sessionStorage;
-  (window as unknown as Record<symbol, unknown>)[Symbol.for("lsi-extension-flag")] = true;
-  storage.setItem(key, value);
-}
-
-function removeFromStorage(storageType: string, key: string): void {
-  const storage = storageType === "localStorage" ? localStorage : sessionStorage;
-  (window as unknown as Record<symbol, unknown>)[Symbol.for("lsi-extension-flag")] = true;
-  storage.removeItem(key);
-}
-
-function importToStorage(storageType: string, entries: Record<string, string>, clearFirst: boolean): void {
-  const storage = storageType === "localStorage" ? localStorage : sessionStorage;
-  const FLAG = Symbol.for("lsi-extension-flag");
-  if (clearFirst) {
-    (window as unknown as Record<symbol, unknown>)[FLAG] = true;
-    storage.clear();
-  }
-  for (const [key, value] of Object.entries(entries)) {
-    (window as unknown as Record<symbol, unknown>)[FLAG] = true;
-    storage.setItem(key, value);
-  }
-}
 
 export function App() {
   const [storageType, setStorageType] = useState<StorageType>("localStorage");
